+++ R/SetMethods.R
@@ -159,33 +159,6 @@ setMethod("filteredCTSSidx<-", "CAGEexp", function (object, value) {
   if (validObject(object)) object
 })
 
-
-#' @name CTSSclusteringMethod<-
-#' 
-#' @rdname CTSSclusteringMethod
-#' 
-#' @param value character
-
-setGeneric( "CTSSclusteringMethod<-"
-          , function(object, value) standardGeneric("CTSSclusteringMethod<-"))
-
-#' @rdname CTSSclusteringMethod
-
-setMethod("CTSSclusteringMethod<-", "GRangesList", function (object, value) {
-  metadata(object)$clusteringMethod <- value
-  if (validObject(object)) object
-})
-
-#' @rdname CTSSclusteringMethod
-
-setMethod("CTSSclusteringMethod<-", "CAGEexp", function (object, value) {
-  CTSSclusteringMethod(metadata(object)$tagClusters) <- value
-  # extrat directly TCs from metadata slot because tagClustersGR does more that
-  # is not needed here.
-  if (validObject(object)) object
-})
-
-
 #' @name  CTSScumulativesTagClusters<-
 #' 
 #' @rdname CTSScumulativesTagClusters